+++ clients/admin-ui/src/types/api/models/RuleTarget.ts
@@ -2,13 +2,11 @@
 /* tslint:disable */
 /* eslint-disable */
 
-import type { FidesDataCategory } from "./FidesDataCategory";
-
 /**
  * An external representation of a Rule's target DataCategory within a Fidesops Policy
  */
 export type RuleTarget = {
   name?: string;
   key?: string;
-  data_category: FidesDataCategory;
+  data_category: string;
 };